Move test codes to their own package
