+++ src/main/java/com/github/crimsondawn45/fabricshieldlib/initializers/FabricShieldLib.java
@@ -2,6 +2,7 @@ package com.github.crimsondawn45.fabricshieldlib.initializers;
 
 import com.github.crimsondawn45.fabricshieldlib.lib.config.FabricShieldLibConfig;
 import com.github.crimsondawn45.fabricshieldlib.lib.object.FabricShieldDecoratorRecipe;
+import com.github.crimsondawn45.fabricshieldlib.tests.FabricShieldLibTests;
 
 import eu.midnightdust.lib.config.MidnightConfig;
 import net.fabricmc.api.ModInitializer;
